batch module authorization

+++ src/modules/Batch/containers/BatchList.js
@@ -7,7 +7,12 @@ import { useNavigate, useLocation } from "react-router-dom";
 import { useBatch } from "../hooks";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
 
-export const BatchList = () => {
+export const BatchList = (props) => {
+  const { hasPermission } = props;
+  const hasCreatePermission = hasPermission("Batches", "Create");
+  const hasEditPermission = hasPermission("Batches", "Edit");
+  const hasDeletePermission = hasPermission("Batches", "Delete");
+
   const { batch, batchesQuery, onDelete, onToggleModal, isConfirmDelete, deleteBatch } = useBatch({
     load: true,
   });
@@ -26,12 +31,17 @@ export const BatchList = () => {
   const ActionButtons = ({ value }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
@@ -68,7 +78,9 @@ export const BatchList = () => {
       subtitle={"List"}
       breadcrumb={[{ label: "Batch", location: "/batch" }]}
     >
-      <BatchFilter />
+      {
+        hasCreatePermission && <BatchFilter />
+      }
       <TableLayout columns={columns} data={data} />
       <ModalLayout
         isOpen={isConfirmDelete}
